+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to `elfinder` will be documented in this file
 
+## 1.0.5 - 2024-01-12
+
+- https://github.com/FriendsOfBotble/elfinder/compare/1.0.4...1.0.5
+- Fix: minor bugs
+- Improve: code styles, apply PSR-12 for code styles
+- Improve: replace the media public path in `RVMedia::class`
+- Feature: add `elfinder:thumbnail:generate` command to generate thumbnails for all files in the media folder
+
+## 1.0.4 - 2024-01-12
+
+- https://github.com/FriendsOfBotble/elfinder/compare/1.0.0...1.0.4
+- Fix: minor bugs
+- Improve code styles
+
+## 1.0.4 - 2024-01-12
+
+- https://github.com/FriendsOfBotble/elfinder/compare/1.0.0...1.0.4
+- Fix: minor bugs
+- Improve code styles
+
 ## 1.0.0 - 2024-01-09
 
 - First release 🥳